Refactor: PhotoInfo 컴포넌트를 PhotocardItem 을 세로로 보여주는 형태로 재작성
