+++ client/src/component/skeleton/SkeletonItem.tsx
@@ -8,12 +8,7 @@ interface Props {
 }
 
 function SkeletonItem({ className, styles }: Props) {
-  return (
-    <StyledItem
-      className={classNames("SkeletonItem", className)}
-      {...styles}
-    />
-  );
+  return <StyledItem className={classNames('SkeletonItem', className)} {...styles} />;
 }
 
 export default SkeletonItem;
@@ -23,6 +18,7 @@ interface StylesProps {
   width?: string;
   maxWidth?: string;
   height?: string;
+  aspectRatio?: string;
   padding?: string;
   margin?: string;
   marginLeft?: string;
@@ -34,16 +30,17 @@ interface StylesProps {
   boxSizing?: string;
 }
 const StyledItem = styled.div<StylesProps>`
-  width: ${p => p.width ? p.width : '5em'};
+  width: ${p => p.width};
   max-width: ${p => p.maxWidth};
-  height: ${p => p.height ? p.height : '1.25em'};
+  height: ${p => p.height};
+  aspect-ratio: ${p => p.aspectRatio};
   padding: ${p => p.padding};
   margin: ${p => p.margin};
   margin-top: ${p => p.marginTop};
   margin-bottom: ${p => p.marginBottom};
   margin-left: ${p => p.marginLeft};
   margin-right: ${p => p.marginRight};
-  background-color: ${p => p.backgroundColor ? p.backgroundColor : 'gainsboro'};
-  border-radius: ${p => p.borderRadius ? p.borderRadius : "10px"};
+  background-color: ${p => p.backgroundColor || 'gainsboro'};
+  border-radius: ${p => p.borderRadius || '10px'};
   box-sizing: ${p => p.boxSizing};
 `;
